NSFRC doesnt fire a required delete then insert
This demonstrates a lot of different changes that can happen in a single
NSManagedObjectContextDidSaveNotification.

NSFRC improperly reports updates as moves, which breaks UITableView. We
can work around this easily by converting a move to an delete and then
re-insert.

However, if an entity E1 is deleted at slot S and then a new entity E2
is inserted at slot S, NSFRC doesn't report a delete for E1 and then an
insert for E2. This fundamentally breaks NSFRC.